backend/authentication/pyproject.toml
[tool.poetry]
name = "aniblog"
version = "0.1.0"
description = ""
authors = ["Dzmitry Zhybryk"]
packages = [
    { include = "app" }
]


[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.81.0"
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"
orm = "^0.3.1"
passlib = "^1.7.4"
databases = "^0.6.1"
asyncpg = "^0.26.0"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
aiosqlite = "^0.17.0"
bcrypt = "^4.0.0"
fastapi-jwt-auth = "^0.5.0"



[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
requests = "^2.28.1"

[tool.poetry.scripts]
dev = "app.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
